autotagger/requirements.txt
de-core-news-lg==3.7.0
fastapi==0.110.0
numpy==1.26.4
rank_bm25==0.2.2
PyPDF2==3.0.1
python-docx==1.1.0
python-multipart==0.0.9
spacy==3.7.4
tqdm==4.66.2
uvicorn==0.29.0